+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+1.8.3
+
+- added: dropdown menu for camera settings with focal length tweak and presets (hardcoded for now)
+- fixed: better display of lens value in camera UI list
+
 1.8.2
 
 - changed: centered toolbar switch